Changes
General 
 
I have decided to make the components GUI as close as possible to the spec provided. It has been tested and work in FireFox, Chrome, internet explorer, and edge. Most of the code has been written in jQuery in order to make it concise and for cross browser compatibility purposes, however the currency converter is coded in JS. The component is also responsive, using twitter bootstrap and media queries. 
 
The JSON data provided remains unedited and I have included a resources object to hold language data used by i18n. I have used the JSON data to access the images, main heading and prices for product items. I have also done this for the product names and URLs, but have combined those with method chaining. 
 
I have used jQuery to change the currency number format and symbol type on button click for each selector as this was more concise than vanilla javascript, where variable values are changed for the desired currency result. 
 
The currency converter function not only decide the cu